+++ src/org/pepstock/charba/client/defaults/globals/DefaultScale.java
@@ -38,11 +38,11 @@ public final class DefaultScale implements IsDefaultScale {
 
 	private static final boolean DEFAULT_OFFSET = false;
 
-	private static final int DEFAULT_WEIGHT = 0;
+	private static final double DEFAULT_WEIGHT = 0D;
 
-	private static final double DEFAULT_BAR_PERCENTAGE = 0.9F;
+	private static final double DEFAULT_BAR_PERCENTAGE = 0.9D;
 
-	private static final double DEFAULT_CATEGORY_PERCENTAGE = 0.8F;
+	private static final double DEFAULT_CATEGORY_PERCENTAGE = 0.8D;
 
 	private static final int DEFAULT_BAR_THICKNESS = 0;
 
@@ -148,7 +148,7 @@ public final class DefaultScale implements IsDefaultScale {
 	 * @see org.pepstock.charba.client.defaults.IsDefaultScale#getWeight()
 	 */
 	@Override
-	public int getWeight() {
+	public double getWeight() {
 		return DEFAULT_WEIGHT;
 	}
